requirements.txt

Flask>=2.3
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
python-dotenv>=1.0